+++ arrus/core/api/devices/us4r/Us4OEM.h
@@ -63,19 +63,6 @@ public:
     * @throws arrus::IllegalStateException when invalid input parameters detected
     */
 	virtual void setAfe(uint8_t address, uint16_t value) = 0;
-    //virtual void setAfeFir(uint8_t address, uint16_t* coeffs, uint8_t length) = 0;
-
-    /**
-    * Enables and configures AFE built-in demodulator
-    * 
-    * @param demodulationFrequency: Demodulation frequency
-    * @param decimationFactor: Decimation factor
-    * @param firCoefficients: Pointer to Low pass filter coefficients buffer
-    * @param nCoefficients: Number of FIR coefficients
-    * @throws arrus::IllegalStateException when invalid input parameters detected
-    */
-    virtual void setAfeDemod(float demodulationFrequency, float decimationFactor, const int16 *firCoefficients,
-                             size_t nCoefficients) = 0;
 
     /**
     * Enables and configures AFE built-in demodulator
@@ -94,7 +81,7 @@ public:
     *
     */
     virtual void disableAfeDemod() = 0;
-	
+
     /**
      * Checks if the firmware version on the Us4OEM module is correct.
      *
@@ -121,6 +108,28 @@ public:
      * Returns Tx component firmware version installed on this us4OEM module.
      */
     virtual uint32 getTxFirmwareVersion() = 0;
+
+    /**
+     * Returns current FPGA wall clock (time passed since Init function was called).
+     *
+     * @return FPGA wall clock (seconds)
+     */
+    virtual float getFPGAWallclock() = 0;
+
+    /**
+     * Enables High-Pass Filter and sets a given corner frequency.
+     *
+     * Available corner frequency values (Hz): 4520'000, 2420'000, 1200'000, 600'000, 300'000, 180'000,
+     * 80'000, 40'000, 20'000.
+     *
+     * @param frequency corner high-pass filter frequency to set
+     */
+    virtual void setHpfCornerFrequency(uint32_t frequency) = 0;
+
+    /**
+     * Disables digital high-pass filter.
+     */
+    virtual void disableHpf() = 0;
 
     Us4OEM(Us4OEM const&) = delete;
     Us4OEM(Us4OEM const&&) = delete;