+++ arrus/core/devices/us4r/Us4RImpl.h
@@ -125,6 +125,11 @@ public:
     float getUCDMeasuredHVMVoltage(uint8_t oemId) override;
     void setStopOnOverflow(bool isStopOnOverflow) override;
     bool isStopOnOverflow() const override;
+    void setHpfCornerFrequency(uint32_t frequency) override;
+    void disableHpf() override;
+
+    uint16_t getAfe(uint8_t reg) override;
+    void setAfe(uint8_t reg, uint16_t val) override;
 
 
 private: